refactor:Entity와 관련된 Integer 모두 Long으로 변경

+++ src/main/java/com/shootdoori/match/dto/TeamDetailResponseDto.java
@@ -10,7 +10,7 @@ public record TeamDetailResponseDto(
     String university,
     SkillLevel skillLevel,
     TeamType teamType,
-    Integer memberCount,
+    Long memberCount,
     String createdAt
 ) {
